requirements.txt
# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7
django-environ==0.11.2

# Async and WebSocket Support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Background Tasks
django-rq==2.8.1
redis==5.0.1

# Web Scraping
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
requests==2.31.0
fake-useragent==1.4.0

# AI and ML
openai==1.3.7
httpx==0.25.2

# Data Processing
pandas==2.1.3
numpy==1.25.2
python-dateutil==2.8.2

# Utilities
python-decouple==3.8
celery==5.3.4
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-debug-toolbar==4.2.0
